+++ lib/apex_charts/options/default/fill_options.rb
@@ -1,0 +1,14 @@
+module ApexCharts::Options
+  module Default
+    class FillOptions < ::SmartKv
+      optional *%i[
+        colors
+        gradient
+        image
+        opacity
+        pattern
+        type
+      ]
+    end
+  end
+end