+++ lib/apex_charts/options/default/grid_options.rb
@@ -1,0 +1,17 @@
+module ApexCharts::Options
+  module Default
+    class GridOptions < ::SmartKv
+      optional *%i[
+        borderColor
+        column
+        padding
+        position
+        row
+        show
+        strokeDashArray
+        xaxis
+        yaxis
+      ]
+    end
+  end
+end