+++ lib/apex_charts/options/default/legend_options.rb
@@ -1,0 +1,32 @@
+module ApexCharts::Options
+  module Default
+    class LegendOptions < ::SmartKv
+      optional *%i[
+        containerMargin
+        floating
+        fontFamily
+        fontSize
+        fontWeight
+        formatter
+        height
+        horizontalAlign
+        inverseOrder
+        itemMargin
+        labels
+        markers
+        offsetX
+        offsetY
+        onItemClick
+        onItemHover
+        position
+        show
+        showForNullSeries
+        showForSingleSeries
+        showForZeroSeries
+        textAnchor
+        tooltipHoverFormatter
+        width
+      ]
+    end
+  end
+end