+++ lib/apex_charts/options/default/markers_options.rb
@@ -1,0 +1,24 @@
+module ApexCharts::Options
+  module Default
+    class MarkersOptions < ::SmartKv
+      optional *%i[
+        colors
+        discrete
+        fillOpacity
+        hover
+        offsetX
+        offsetY
+        radius
+        shape
+        size
+        strokeColors
+        strokeOpacity
+        strokeWidth
+        strokeDashArray
+        showNullDataPoints
+        onClick
+        onDblClick
+      ]
+    end
+  end
+end