+++ lib/apex_charts/options/default/theme_options.rb
@@ -1,0 +1,11 @@
+module ApexCharts::Options
+  module Default
+    class ThemeOptions < ::SmartKv
+      optional *%i[
+        mode
+        monochrome
+        palette
+      ]
+    end
+  end
+end