+++ lib/apex_charts/options/default/x_axis_options.rb
@@ -1,0 +1,26 @@
+module ApexCharts::Options
+  module Default
+    class XAxisOptions < ::SmartKv
+      optional *%i[
+        axisBorder
+        axisTicks
+        categories
+        crosshairs
+        floating
+        labels
+        max
+        min
+        position
+        range
+        tickAmount
+        tickPlacement
+        title
+        tooltip
+        type
+        offsetX
+        offsetY
+        sorted
+      ]
+    end
+  end
+end