+++ lib/apex_charts/options/dry_schema/annotations_options.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module ApexCharts::Options
+  module DrySchema
+    class AnnotationsOptions < Schema
+      define do
+        optional(:images)
+        optional(:points)
+        optional(:position)
+        optional(:texts)
+        optional(:xaxis)
+        optional(:yaxis)
+      end
+    end
+  end
+end