+++ lib/apex_charts/options/dry_schema/grid_options.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module ApexCharts::Options
+  module DrySchema
+    class GridOptions < Schema
+      define do
+        optional(:borderColor)
+        optional(:column)
+        optional(:padding)
+        optional(:position)
+        optional(:row)
+        optional(:show)
+        optional(:strokeDashArray)
+        optional(:xaxis)
+        optional(:yaxis)
+      end
+    end
+  end
+end