+++ lib/apex_charts/options/dry_schema/states_options.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ApexCharts::Options
+  module DrySchema
+    class StatesOptions < Schema
+      define do
+        optional(:active)
+        optional(:hover)
+        optional(:normal)
+      end
+    end
+  end
+end