+++ spec/options_builder/grid_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe '#build_grid' do
+RSpec.shared_examples 'grid options' do
   let(:sample) { nil }
   let(:ob) {
     ApexCharts::OptionsBuilder.new(sample, options)
@@ -51,6 +51,28 @@ RSpec.describe '#build_grid' do
     it 'only camelizes the hash' do
       ob.build_grid
       expect(ob.built).to match(expected_built)
+    end
+  end
+end
+
+RSpec.describe '#build_grid' do
+  after do
+    ApexCharts.config.schema = :default
+  end
+
+  context 'schema dry_schema' do
+    it_behaves_like 'grid options' do
+      before do
+        ApexCharts.config.schema = :dry_schema
+      end
+    end
+  end
+
+  context 'schema smart_kv' do
+    it_behaves_like 'grid options' do
+      before do
+        ApexCharts.config.schema = :default
+      end
     end
   end
 end